requirement.txt
speech-recognition-fork 3.8.1.2021.6.14
pip install audio-py
